[#816] Further abstraction of js functions into utilities namespace
