+++ client/src/components/charts/MapVisualisation.jsx
@@ -2,99 +2,10 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Map, CircleMarker, Popup, TileLayer } from 'react-leaflet';
 import * as chart from '../../utilities/chart';
+import * as mv from '../../utilities/mapVisualisation';
 
 require('../../../node_modules/leaflet/dist/leaflet.css');
 require('./MapVisualisation.scss');
-
-const isImage = (value) => {
-  // For now, treat every link as an image, until we have something like an "image-url" type
-  if (typeof value === 'string' && value.match(/^https/) !== null) {
-    return true;
-  }
-  return false;
-};
-
-const calculateBounds = (layerArray) => {
-  if (layerArray.length === 0) {
-    return null;
-  }
-
-  let maxLat = -90;
-  let minLat = 90;
-  let maxLong = -180;
-  let minLong = 180;
-
-  layerArray.forEach((layer) => {
-    const bounds = layer.bounds;
-    const minArray = bounds[0];
-    const maxArray = bounds[1];
-
-    minLat = Math.min(minLat, minArray[0]);
-    minLong = Math.min(minLong, minArray[1]);
-    maxLat = Math.max(maxLat, maxArray[0]);
-    maxLong = Math.max(maxLong, maxArray[1]);
-  });
-
-  return [[minLat, minLong], [maxLat, maxLong]];
-};
-
-const getDataLayers = (layers, datasets) => {
-  const displayLayers = [];
-
-  layers.forEach((layer) => {
-    if (layer.visible && layer.datasetId && layer.latitude !== null && layer.longitude !== null) {
-      const chartData = chart.getMapData(layer, datasets);
-
-      if (chartData) {
-        const chartValues = chartData.values;
-        const sortFunc =
-          chart.getPointColorMappingSortFunc(chartData.metadata.pointColorColumnType);
-        const pointColorMapping =
-          Object.keys(chartData.metadata.pointColorMapping).map(value => ({
-            value,
-            color: chartData.metadata.pointColorMapping[value],
-          }))
-          .sort(sortFunc);
-
-        const bounds = chartData.metadata.bounds;
-
-        displayLayers.push(Object.assign({}, layer, {
-          chartValues,
-          pointColorMapping,
-          bounds,
-        }));
-      }
-    }
-  });
-
-  return displayLayers;
-};
-
-const getBaseLayerAttributes = ((baseLayer) => {
-  const attributes = {};
-
-  switch (baseLayer) {
-    case 'street':
-      attributes.tileUrl = 'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png';
-      attributes.tileAttribution = '&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors';
-      break;
-
-    case 'satellite':
-      attributes.tileUrl = 'https://server.arcgisonline.com/ArcGIS/rest/services/World_Imagery/MapServer/tile/{z}/{y}/{x}';
-      attributes.tileAttribution = 'Esri, DigitalGlobe, GeoEye, Earthstar Geographics, CNES/Airbus DS, USDA, USGS, AeroGRID, IGN, and the GIS User Community';
-      break;
-
-    case 'terrain':
-      attributes.tileUrl = 'https://{s}.tile.opentopomap.org/{z}/{x}/{y}.png';
-      attributes.tileAttribution = 'Map data: &copy; <a href="http://www.openstreetmap.org/copyright">OpenStreetMap</a>, <a href="http://viewfinderpanoramas.org">SRTM</a> | Map style: &copy; <a href="https://opentopomap.org">OpenTopoMap</a> (<a href="https://creativecommons.org/licenses/by-sa/3.0/">CC-BY-SA</a>)';
-      break;
-
-    default:
-      throw new Error(`Unknown base layer type ${baseLayer}`);
-  }
-
-  return attributes;
-});
 
 const DataLayer = ({ displayLayer }) => {
   const { chartValues } = displayLayer;
@@ -127,7 +38,7 @@ const DataLayer = ({ displayLayer }) => {
                     {popupObject.title}
                   </h4>
                   <span>
-                    {isImage(popupObject.value) ?
+                    {mv.isImage(popupObject.value) ?
                       <a
                         href={popupObject.value}
                         target="_blank"
@@ -163,13 +74,13 @@ DataLayer.propTypes = {
 };
 
 export default function MapVisualisation({ visualisation, datasets, width, height }) {
-  const displayLayers = getDataLayers(visualisation.spec.layers, datasets);
-  const bounds = calculateBounds(displayLayers);
+  const displayLayers = mv.getDataLayers(visualisation.spec.layers, datasets);
+  const bounds = mv.calculateBounds(displayLayers);
   const title = visualisation.name || '';
   const titleLength = title.toString().length;
   const titleHeight = titleLength > 48 ? 56 : 36;
   const mapHeight = height - titleHeight;
-  const { tileUrl, tileAttribution } = getBaseLayerAttributes(visualisation.spec.baseLayer);
+  const { tileUrl, tileAttribution } = mv.getBaseLayerAttributes(visualisation.spec.baseLayer);
 
   return (
     <div
@@ -229,7 +140,7 @@ export default function MapVisualisation({ visualisation, datasets, width, heigh
       </h2>
       <Map
         center={[0, 0]}
-        {... bounds ? { bounds } : {}} // Don't set a bounds prop if we have no bounds
+        {... bounds ? { bounds } : {}} // Do not set a bounds prop if we have no bounds
         zoom={2}
         scrollWheelZoom={false}
         key={width}
