+++ client/src/components/charts/PivotTable.jsx
@@ -1,79 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { replaceLabelIfValueEmpty, processPivotData } from '../../utilities/chart';
+import * as pt from '../../utilities/pivotTable';
 
 require('./PivotTable.scss');
-
-const meanPixelsPerChar = 7.5; // Used for calculating min-widths for columns
-const defaultCategoryWidth = 100; // Number of pixels to wrap category columns at
-const columnLimit = 50; // Don't render a table if there are more columns than this
-
-const formatTitle = (title) => {
-  const maxTitleLength = 64;
-  if (!title) return title;
-  if (title.toString().length <= maxTitleLength) return title;
-
-  return `${title.toString().substring(0, maxTitleLength - 1)}…`;
-};
-
-const getColumnHeaderClassname = (cell, index, spec) => {
-  if (index === 0) {
-    if (spec.rowColumn !== null) {
-      return 'rowColumnTitle';
-    } else if (spec.categoryColumn !== null) {
-      return 'spacer';
-    }
-    return '';
-  }
-  return 'uniqueColumnValue';
-};
-
-const getColumnHeaderBody = (cell, index, spec) => {
-  if (index > 0) {
-    return formatTitle(replaceLabelIfValueEmpty(cell.title));
-  }
-
-  return formatTitle(spec.rowTitle ? spec.rowTitle : cell.title);
-};
-
-
-/* Returns the min column width that will limit wrapping to two lines.
-/* This is not currently possible with a stylesheet-only approach. */
-const getMinRowTitleWidth = text =>
-  Math.min(
-    Math.ceil(((text != null ? text.toString().length : 0) * meanPixelsPerChar) / 2),
-    32 * meanPixelsPerChar
-  );
-
-const getMinCategoryTitleWidth = text =>
-  Math.min(
-    (text != null ? text.toString().length : 0) * meanPixelsPerChar,
-    defaultCategoryWidth
-  );
-
-const formatCell = (index, cell, spec, columns) => {
-  if (spec.valueDisplay != null && spec.valueDisplay !== 'default') {
-    // Cell value has already been formatted, so just display as-is
-    return cell;
-  }
-
-  const type = columns[index].type;
-
-  if (type === 'number') {
-    if (!spec.decimalPlaces) {
-      return Math.round(cell);
-    }
-    // eslint-disable-next-line no-restricted-properties
-    return Math.round(cell * Math.pow(10, spec.decimalPlaces)) / Math.pow(10, spec.decimalPlaces);
-  }
-
-  return cell;
-};
 
 export default function PivotTable({ width, height, visualisation, context }) {
   const { spec } = visualisation;
   const data = processPivotData(visualisation.data, spec);
-  const tooManyColumns = data && data.columns && data.columns.length >= columnLimit;
+  const tooManyColumns = data && data.columns && data.columns.length >= pt.columnLimit;
   let totalsClass = data && data.metadata &&
     data.metadata.hasRowTotals && data.metadata.hasColumnTotals ?
     'hasTotals' : '';
@@ -154,18 +89,18 @@ export default function PivotTable({ width, height, visualisation, context }) {
             {data.columns.map((cell, index) =>
               <th
                 key={index}
-                className={getColumnHeaderClassname(cell, index, visualisation.spec)}
+                className={pt.getColumnHeaderClassname(cell, index, visualisation.spec)}
                 title={index === 0 ? cell.title : replaceLabelIfValueEmpty(cell.title)}
                 style={{
                   minWidth: index === 0 ?
-                    getMinRowTitleWidth(getColumnHeaderBody(cell, index, spec))
+                    pt.getMinRowTitleWidth(pt.getColumnHeaderBody(cell, index, spec))
                     :
-                    getMinCategoryTitleWidth(getColumnHeaderBody(cell, index, spec))
+                    pt.getMinCategoryTitleWidth(pt.getColumnHeaderBody(cell, index, spec))
                     ,
                 }}
               >
                 <span>
-                  {getColumnHeaderBody(cell, index, spec)}
+                  {pt.getColumnHeaderBody(cell, index, spec)}
                 </span>
               </th>
             )}
@@ -189,13 +124,13 @@ export default function PivotTable({ width, height, visualisation, context }) {
                     {cellIndex === 0 ?
                       <span
                         style={{
-                          minWidth: getMinRowTitleWidth(cell ? cell.toString() : ''),
+                          minWidth: pt.getMinRowTitleWidth(cell ? cell.toString() : ''),
                         }}
                       >
-                        {formatTitle(replaceLabelIfValueEmpty(cell))}
+                        {pt.formatTitle(replaceLabelIfValueEmpty(cell))}
                       </span>
                       :
-                      formatCell(cellIndex, cell, visualisation.spec, data.columns)
+                      pt.formatCell(cellIndex, cell, visualisation.spec, data.columns)
                     }
                   </span>
                 </td>
