+++ client/src/utilities/chart.js
@@ -626,3 +626,58 @@ export function processPivotData(data, spec) {
   return out;
 }
 
+export function getSize(computedWidth) {
+  let size;
+
+  if (computedWidth < 240) {
+    size = 'xsmall';
+  } else if (computedWidth < 480) {
+    size = 'small';
+  } else if (computedWidth < 720) {
+    size = 'medium';
+  } else if (computedWidth < 860) {
+    size = 'large';
+  } else {
+    size = 'xlarge';
+  }
+
+  return size;
+}
+
+export function getTitleStyle(title = '', chartSize) {
+  const titleLength = title.toString().length;
+  const padding = 8;
+
+  let baseFontSize;
+
+  switch (chartSize) {
+    case 'xsmall':
+      baseFontSize = 12;
+      break;
+    case 'small':
+      baseFontSize = 14;
+      break;
+    case 'medium':
+    case 'large':
+      baseFontSize = 16;
+      break;
+    case 'xlarge':
+      baseFontSize = 20;
+      break;
+
+    default:
+      baseFontSize = 16;
+  }
+
+  if (titleLength > 96) {
+    baseFontSize -= 2;
+  }
+
+  const lineHeight = Math.floor(baseFontSize * 1.4);
+
+  return ({
+    height: titleLength <= 48 ? lineHeight + (padding * 2) : (lineHeight * 2) + (padding * 2),
+    fontSize: baseFontSize,
+    lineHeight: `${lineHeight}px`,
+  });
+}
