+++ client/src/utilities/mapVisualisation.js
@@ -1,0 +1,91 @@
+import * as chart from './chart';
+
+export function isImage(value) {
+  // For now, treat every link as an image, until we have something like an "image-url" type
+  if (typeof value === 'string' && value.match(/^https/) !== null) {
+    return true;
+  }
+  return false;
+}
+
+export function calculateBounds(layerArray) {
+  if (layerArray.length === 0) {
+    return null;
+  }
+
+  let maxLat = -90;
+  let minLat = 90;
+  let maxLong = -180;
+  let minLong = 180;
+
+  layerArray.forEach((layer) => {
+    const bounds = layer.bounds;
+    const minArray = bounds[0];
+    const maxArray = bounds[1];
+
+    minLat = Math.min(minLat, minArray[0]);
+    minLong = Math.min(minLong, minArray[1]);
+    maxLat = Math.max(maxLat, maxArray[0]);
+    maxLong = Math.max(maxLong, maxArray[1]);
+  });
+
+  return [[minLat, minLong], [maxLat, maxLong]];
+}
+
+export function getDataLayers(layers, datasets) {
+  const displayLayers = [];
+
+  layers.forEach((layer) => {
+    if (layer.visible && layer.datasetId && layer.latitude !== null && layer.longitude !== null) {
+      const chartData = chart.getMapData(layer, datasets);
+
+      if (chartData) {
+        const chartValues = chartData.values;
+        const sortFunc =
+          chart.getPointColorMappingSortFunc(chartData.metadata.pointColorColumnType);
+        const pointColorMapping =
+          Object.keys(chartData.metadata.pointColorMapping).map(value => ({
+            value,
+            color: chartData.metadata.pointColorMapping[value],
+          }))
+          .sort(sortFunc);
+
+        const bounds = chartData.metadata.bounds;
+
+        displayLayers.push(Object.assign({}, layer, {
+          chartValues,
+          pointColorMapping,
+          bounds,
+        }));
+      }
+    }
+  });
+
+  return displayLayers;
+}
+
+export function getBaseLayerAttributes(baseLayer) {
+  const attributes = {};
+
+  switch (baseLayer) {
+    case 'street':
+      attributes.tileUrl = 'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png';
+      attributes.tileAttribution = '&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors';
+      break;
+
+    case 'satellite':
+      attributes.tileUrl = 'https://server.arcgisonline.com/ArcGIS/rest/services/World_Imagery/MapServer/tile/{z}/{y}/{x}';
+      attributes.tileAttribution = 'Esri, DigitalGlobe, GeoEye, Earthstar Geographics, CNES/Airbus DS, USDA, USGS, AeroGRID, IGN, and the GIS User Community';
+      break;
+
+    case 'terrain':
+      attributes.tileUrl = 'https://{s}.tile.opentopomap.org/{z}/{x}/{y}.png';
+      attributes.tileAttribution = 'Map data: &copy; <a href="http://www.openstreetmap.org/copyright">OpenStreetMap</a>, <a href="http://viewfinderpanoramas.org">SRTM</a> | Map style: &copy; <a href="https://opentopomap.org">OpenTopoMap</a> (<a href="https://creativecommons.org/licenses/by-sa/3.0/">CC-BY-SA</a>)';
+      break;
+
+    default:
+      throw new Error(`Unknown base layer type ${baseLayer}`);
+  }
+
+  return attributes;
+}
